Rename contact-history feature to contacts
This feature’s scope will expand to be a contacts directory as well
