--- plugin-flex-ts-template-v2/src/feature-library/contact-history/types/ServiceConfiguration.ts
@@ -1,4 +1,0 @@
-export default interface ContactHistoryConfig {
-  enabled: boolean;
-  max_contacts: number;
-}
