+++ plugin-flex-ts-template-v2/src/feature-library/contacts/custom-components/ContactRecord.tsx
