+++ plugin-flex-ts-template-v2/src/feature-library/contacts/flex-hooks/strings/es-mx.json
