+++ plugin-flex-ts-template-v2/src/feature-library/contacts/index.ts
@@ -5,5 +5,5 @@ import hooks from './flex-hooks/**/*.*';
 
 export const register = (): FeatureDefinition => {
   if (!isFeatureEnabled()) return {};
-  return { name: 'contact-history', hooks: typeof hooks === 'undefined' ? [] : hooks };
+  return { name: 'contacts', hooks: typeof hooks === 'undefined' ? [] : hooks };
 };
