+++ plugin-flex-ts-template-v2/src/feature-library/contacts/types/ServiceConfiguration.ts
@@ -1,0 +1,4 @@
+export default interface ContactsConfig {
+  enabled: boolean;
+  max_contacts: number;
+}
